add test case & sample
