+++ examples/point.less
@@ -1,0 +1,3 @@
+.point-popup {
+  pointer-events: none;
+}
