+++ tests/index.less
@@ -1,0 +1,5 @@
+// Clean up
+html, body {
+  margin: 0!important;
+  padding: 0!important;
+}
