+++ tests/point.less
@@ -1,0 +1,10 @@
+.point-region {
+  width: 1128px;
+  height: 903px;
+}
+
+.point-popup {
+  width: 100px;
+  height: 50px;
+  white-space: nowrap;
+}
